requirements.txt
addict
future
lmdb
pyyaml
requests
scikit-image
scipy==1.10.1
tb-nightly
torch>=1.7.1
torchvision>=0.8.2
tqdm
yapf
lpips
gdown  # supports downloading the large file from Google Drive
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
basicsr==1.4.2
facexlib==0.3.0
realesrgan==0.3.0
gfpgan==1.3.8